+++ app/decorators/pathfinder2_character/base_decorator.rb
@@ -10,6 +10,7 @@ module Pathfinder2Character
              :lore_skills, :background, :weapon_skills, :armor_skills, :main_ability, :class_dc,
              :saving_throws, :dying_condition_value, :ability_boosts, :skill_boosts, :coins, to: :data
 
+    def parent = __getobj__
     def method_missing(_method, *args); end
 
     def boosts